backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.1
pydantic==2.7.0
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
email-validator==2.1.1
